+++ src/core/watchQueryOptions.ts
@@ -1,4 +1,4 @@
-import { DocumentNode, ExecutionResult } from 'graphql';
+import { DocumentNode } from 'graphql';
 
 import { ApolloCache } from '../cache/core/cache';
 import { FetchResult } from '../link/core/types';
@@ -53,11 +53,6 @@ export interface QueryBaseOptions<TVariables = OperationVariables> {
    * Specifies the {@link ErrorPolicy} to be used for this query
    */
   errorPolicy?: ErrorPolicy;
-
-  /**
-   * Whether or not to fetch results
-   */
-  fetchResults?: boolean;
 
   /**
    * Context to be passed to link execution chain
@@ -198,7 +193,7 @@ export interface MutationBaseOptions<
    * once these queries return.
    */
   refetchQueries?:
-    | ((result: ExecutionResult<T>) => RefetchQueryDescription)
+    | ((result: FetchResult<T>) => RefetchQueryDescription)
     | RefetchQueryDescription;
 
   /**